Anagram detection algorithm
Observations: 
Given two strings of characters, determine if they are anagrams. A word is an anagram of another if it contains exactly the same letters. For example, "mora" and "roma".

This problem has been solved considering  2 words as given arguments in the call to the function "anagram_dectector".  It can detect when the arguments are not string types or contains numeric characters.